+++ pkgs/ffigen/lib/src/code_generator/objc_built_in_functions.dart
@@ -42,7 +42,7 @@ class ObjCBuiltInFunctions {
       ObjCImport('UnimplementedOptionalMethodException');
 
   // Keep in sync with pkgs/objective_c/ffigen_objc.yaml.
-  static const builtInInterfaces = {
+  static const _builtInInterfaces = {
     'DartProxy',
     'DartProxyBuilder',
     'NSArray',
@@ -77,11 +77,11 @@ class ObjCBuiltInFunctions {
     'NSValue',
     'Protocol',
   };
-  static const builtInCompounds = {
-    'NSFastEnumerationState',
-    'NSRange',
+  static const _builtInCompounds = {
+    'NSFastEnumerationState': 'NSFastEnumerationState',
+    '_NSRange': 'NSRange',
   };
-  static const builtInEnums = {
+  static const _builtInEnums = {
     'NSBinarySearchingOptions',
     'NSComparisonResult',
     'NSDataBase64DecodingOptions',
@@ -109,11 +109,11 @@ class ObjCBuiltInFunctions {
   // TODO(https://github.com/dart-lang/native/issues/1173): Ideally this check
   // would be based on more than just the name.
   bool isBuiltInInterface(String name) =>
-      !generateForPackageObjectiveC && builtInInterfaces.contains(name);
-  bool isBuiltInCompound(String name) =>
-      !generateForPackageObjectiveC && builtInCompounds.contains(name);
+      !generateForPackageObjectiveC && _builtInInterfaces.contains(name);
+  String? getBuiltInCompoundName(String name) =>
+      generateForPackageObjectiveC ? null : _builtInCompounds[name];
   bool isBuiltInEnum(String name) =>
-      !generateForPackageObjectiveC && builtInEnums.contains(name);
+      !generateForPackageObjectiveC && _builtInEnums.contains(name);
   bool isNSObject(String name) => name == 'NSObject';
 
   // We need to load a separate instance of objc_msgSend for each signature. If